+++ pkg/api/api.go
@@ -105,6 +105,7 @@ func (s *Server) registerAPIs() {
 	group.Entries = append(group.Entries, s.aboutAPIEntries()...)
 	group.Entries = append(group.Entries, s.customDataAPIEntries()...)
 	group.Entries = append(group.Entries, s.profileAPIEntries()...)
+	group.Entries = append(group.Entries, s.prometheusMetricsAPIEntries()...)
 
 	for _, fn := range appendAddonAPIs {
 		fn(s, group)