+++ pkg/object/httpserver/httpserver_test.go
@@ -19,11 +19,13 @@ package httpserver
 
 import (
 	"os"
+	"sync"
 	"testing"
 	"time"
 
 	"github.com/megaease/easegress/pkg/context/contexttest"
 	"github.com/megaease/easegress/pkg/logger"
+	"github.com/megaease/easegress/pkg/option"
 	"github.com/megaease/easegress/pkg/supervisor"
 	"github.com/stretchr/testify/assert"
 )
@@ -44,7 +46,9 @@ port: 38081
 keepAlive: true
 https: false
 `
-	superSpec, err := supervisor.NewSpec(yamlConfig)
+	super := supervisor.NewMock(option.New(), nil, sync.Map{}, sync.Map{}, nil,
+		nil, false, nil, nil)
+	superSpec, err := super.NewSpec(yamlConfig)
 	assert.NoError(err)
 
 	svr := &HTTPServer{}
@@ -57,7 +61,7 @@ port: 38082
 keepAlive: true
 https: false
 `
-	superSpec, err = supervisor.NewSpec(yamlConfig)
+	superSpec, err = super.NewSpec(yamlConfig)
 	assert.NoError(err)
 	svr2 := &HTTPServer{}
 	svr2.Inherit(superSpec, svr, &contexttest.MockedMuxMapper{})